requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-modeladmin-reorder==0.3.1
django-rest-auth==0.9.3
djangorestframework-jwt
mysqlclient==2.0.3
django-versatileimagefield==2.2
gunicorn==20.1.0
boto3==1.17.58
django-storages==1.11.1
django-crispy-forms==1.14.0
django-model-utils==4.0
django-prometheus==2.2.0
django-cors-headers==3.14.0
gevent
greenlet
bcrypt==4.0.1
django-heartbeat==2.2.0
django-crum==0.7.9
logzero==1.7.0
requests
CurrencyConverter==0.17.30
